Move store fakes to folder service test
